+++ packages/common/src/repo/mst/index.ts
@@ -1,2 +1,4 @@
 export * from './mst'
 export * from './diff'
+export * from './walker'
+export * as mstUtil from './util'
